Added Functionality for adding money to the account of a user. Added Functionality for removing money to the account of a user. Implemented Profile Controller. Implemented Profile template directory to store tamplates to user for user profile implementation.

+++ src/main/java/com/PickmeUP/project/model/Account.java
@@ -51,11 +51,11 @@ public class Account {
         return this.creation;
     }
 
-    public void plusBalance(int plus){
+    public void plusBalance(Double plus){
         this.balance= this.balance + plus;
     }
 
-    public void minusBalance(int minus){
+    public void minusBalance(Double minus){
         this.balance = this.balance - minus;
     }
 }
